Refactor main session/signaling IPC into focused modules
